requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
Flask-CORS==4.0.0
Werkzeug==3.0.1
MarkupSafe==2.1.3
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
blinker==1.7.0
protobuf==4.21.12